+++ src/components/AddTransactionForm/methodsType.ts
@@ -1,0 +1,6 @@
+import { UseFormReturn } from 'react-hook-form';
+import { AddTransactionFormFields } from './AddTransactionForm';
+
+export type MethodsType = {
+  methods: UseFormReturn<AddTransactionFormFields>;
+};
